healthcare-doc-generato/healthcare-doc-generato/backend/requirements.txt
# Backend Dependencies
Flask
Flask-CORS
python-dotenv
google-generativeai
pypdf
pytesseract
Pillow
python-docx
requests
